Rydder i forvaltere controller, sletter et par metoder og gjør satskjøring endepunktet mer generell
