+++ src/main/kotlin/no/nav/familie/ba/sak/internal/ForvalterService.kt
@@ -1,13 +1,24 @@
 package no.nav.familie.ba.sak.internal
 
 import no.nav.familie.ba.sak.common.Feil
+import no.nav.familie.ba.sak.config.TaskRepositoryWrapper
 import no.nav.familie.ba.sak.integrasjoner.økonomi.AndelTilkjentYtelseForIverksettingFactory
 import no.nav.familie.ba.sak.integrasjoner.økonomi.ØkonomiService
+import no.nav.familie.ba.sak.kjerne.autovedtak.AutovedtakService
 import no.nav.familie.ba.sak.kjerne.behandling.BehandlingHentOgPersisterService
+import no.nav.familie.ba.sak.kjerne.behandling.BehandlingService
+import no.nav.familie.ba.sak.kjerne.behandling.NyBehandling
 import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandling
+import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingStatus
+import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingType
+import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingÅrsak
 import no.nav.familie.ba.sak.kjerne.beregning.BeregningService
 import no.nav.familie.ba.sak.kjerne.endretutbetaling.EndretUtbetalingAndelService
+import no.nav.familie.ba.sak.kjerne.fagsak.FagsakService
+import no.nav.familie.ba.sak.kjerne.steg.StegService
 import no.nav.familie.ba.sak.kjerne.vedtak.VedtakService
+import no.nav.familie.ba.sak.sikkerhet.SikkerhetContext
+import no.nav.familie.ba.sak.task.IverksettMotOppdragTask
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Propagation
 import org.springframework.transaction.annotation.Transactional
@@ -19,6 +30,11 @@ class ForvalterService(
     private val beregningService: BeregningService,
     private val behandlingHentOgPersisterService: BehandlingHentOgPersisterService,
     private val endretUtbetalingAndelService: EndretUtbetalingAndelService,
+    private val stegService: StegService,
+    private val fagsakService: FagsakService,
+    private val behandlingService: BehandlingService,
+    private val taskRepository: TaskRepositoryWrapper,
+    private val autovedtakService: AutovedtakService,
 ) {
 
     @Transactional
@@ -54,5 +70,32 @@ class ForvalterService(
             behandling = sisteVedtatteBehandling,
             forrigeBehandling = nestSisteVedtatteBehandling,
         )
+    }
+
+    @Transactional(propagation = Propagation.REQUIRES_NEW)
+    fun kjørForenkletSatsendringFor(fagsakId: Long) {
+        val fagsak = fagsakService.hentPåFagsakId(fagsakId)
+
+        val nyBehandling = stegService.håndterNyBehandling(
+            NyBehandling(
+                behandlingType = BehandlingType.REVURDERING,
+                behandlingÅrsak = BehandlingÅrsak.SATSENDRING,
+                søkersIdent = fagsak.aktør.aktivFødselsnummer(),
+                skalBehandlesAutomatisk = true,
+                fagsakId = fagsakId,
+            ),
+        )
+
+        val behandlingEtterVilkårsvurdering =
+            stegService.håndterVilkårsvurdering(nyBehandling)
+
+        val opprettetVedtak =
+            autovedtakService.opprettToTrinnskontrollOgVedtaksbrevForAutomatiskBehandling(
+                behandlingEtterVilkårsvurdering,
+            )
+        behandlingService.oppdaterStatusPåBehandling(nyBehandling.id, BehandlingStatus.IVERKSETTER_VEDTAK)
+        val task =
+            IverksettMotOppdragTask.opprettTask(nyBehandling, opprettetVedtak, SikkerhetContext.hentSaksbehandler())
+        taskRepository.save(task)
     }
 }
